Updated more files and started UI implementation for charge counter

+++ HellkiteCode/Cards/Common/Skills/CooledCore.cs
@@ -15,7 +15,7 @@ public sealed class CooledCore() : HellkiteCard(0, CardType.Skill, CardRarity.Un
     
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
-        int spent = await HellkiteCmd.SpendUpToCharge(Owner.Creature, (int)DynamicVars["MaxCharge"].BaseValue, choiceContext);
+        decimal spent = await HellkiteCmd.SpendUpToCharge(Owner.Creature, (int)DynamicVars["MaxCharge"].BaseValue, choiceContext);
         if (spent > 0) await PowerCmd.Apply<PlatingPower>(choiceContext, Owner.Creature, spent, Owner.Creature, this);
         await CardPileCmd.Draw(choiceContext, DynamicVars.Cards.IntValue, Owner);
     }
